requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-embed-video==1.4.0
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
jsmin==2.2.2
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.4
wrapt==1.12.1
